% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_model_info.R
\name{get_model_info}
\alias{get_model_info}
\title{Retrieve Model Information from the pai_model_list}
\usage{
get_model_info(method)
}
\description{
Retrieve Model Information from the pai_model_list
}
\keyword{internal}
