% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAI-package.R
\docType{package}
\name{mapAI-package}
\alias{mapAI}
\alias{mapAI-package}
\title{mapAI: An R Package for Positional Accuracy Improvement of Historical Maps}
\description{
An R package for Positional Accuracy Improvement of Historical Maps that uses supervised learning algorithms. It uses as input a csv file with homologous points and a vector map of the area of interest in order to correct it.
}
\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com} (\href{https://orcid.org/0000-0001-6387-8791}{ORCID})

}
\keyword{internal}
