% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pai_model.R
\name{plot.pai_model}
\alias{plot.pai_model}
\title{Plot Method for pai_model Objects}
\usage{
\method{plot}{pai_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pai_model}.}

\item{...}{Additional arguments (not used).}
}
\description{
Plot Method for pai_model Objects
}
\details{
The plot method visualizes the model components based on whether the
model is univariate or bivariate.
}
\examples{
# See ?train_pai_model for a complete, runnable example.
}
