% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcps.R
\name{print.gcp}
\alias{print.gcp}
\title{Print Method for gcp Objects}
\usage{
\method{print}{gcp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gcp}.}

\item{...}{Additional arguments (not used).}
}
\description{
Print Method for gcp Objects
}
