% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmert.R
\name{print.helmert}
\alias{print.helmert}
\title{Print a Helmert Model Object}
\usage{
\method{print}{helmert}(x, ...)
}
\arguments{
\item{x}{An object of class \code{helmert}.}

\item{...}{Additional arguments (not used).}
}
\description{
S3 print method for objects of class \code{helmert}.
}
