% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tin_methods.R
\name{print.hybrid_affine_tin_fit}
\alias{print.hybrid_affine_tin_fit}
\title{Print Method for hybrid_affine_tin_fit Objects}
\usage{
\method{print}{hybrid_affine_tin_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{hybrid_affine_tin_fit}.}

\item{...}{Additional arguments (not used).}
}
\description{
Print Method for hybrid_affine_tin_fit Objects
}
