% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_pai_model.R
\name{print.pai_assessment}
\alias{print.pai_assessment}
\title{Print Method for pai_assessment Objects}
\usage{
\method{print}{pai_assessment}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pai_assessment}.}

\item{...}{Additional arguments (not used).}
}
\description{
Provides a concise and formatted summary of the model assessment
results.
}
\examples{
# Assuming `assessment_results` is an object from assess_pai_model()
# print(assessment_results)

}
