% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pai_model.R
\name{print.pai_model}
\alias{print.pai_model}
\title{Print Method for pai_model Objects}
\usage{
\method{print}{pai_model}(x, ...)
}
\arguments{
\item{x}{An object of class \code{pai_model}.}

\item{...}{Additional arguments (not used).}
}
\description{
Print Method for pai_model Objects
}
\details{
The print method provides a concise summary of the trained model,
}
\examples{
# See ?train_pai_model for a complete, runnable example.
}
