% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tin_methods.R
\name{print.summary.hybrid_helmert_tin_fit}
\alias{print.summary.hybrid_helmert_tin_fit}
\title{Print Method for summary.hybrid_helmert_tin_fit Objects}
\usage{
\method{print}{summary.hybrid_helmert_tin_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.hybrid_helmert_tin_fit}.}

\item{...}{Additional arguments (not used).}
}
\description{
Print Method for summary.hybrid_helmert_tin_fit Objects
}
