% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tin_methods.R
\name{print.summary.tin_linear_fit}
\alias{print.summary.tin_linear_fit}
\title{Print Method for summary.tin_linear_fit Objects}
\usage{
\method{print}{summary.tin_linear_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.tin_linear_fit}.}

\item{...}{Additional arguments (not used).}
}
\description{
Print Method for summary.tin_linear_fit Objects
}
