% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tin_methods.R
\name{print.tin_akima_fit}
\alias{print.tin_akima_fit}
\title{Print Method for tin_akima_fit Objects}
\usage{
\method{print}{tin_akima_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{tin_akima_fit}.}

\item{...}{Additional arguments (not used).}
}
\description{
Print Method for tin_akima_fit Objects
}
