% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcps.R
\name{st_crs.gcp}
\alias{st_crs.gcp}
\title{Get Coordinate Reference System of a GCP Object}
\usage{
\method{st_crs}{gcp}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gcp}.}

\item{...}{Additional arguments (not used).}
}
\value{
A \code{crs} object or \code{NA}.
}
\description{
Get Coordinate Reference System of a GCP Object
}
