% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcps.R
\name{summary.gcp}
\alias{summary.gcp}
\title{Summary Method for gcp Objects}
\usage{
\method{summary}{gcp}(object, ...)
}
\arguments{
\item{object}{An object of class \code{gcp}.}

\item{...}{Additional arguments (not used).}
}
\description{
Summary Method for gcp Objects
}
\details{
The summary method provides a concise overview of the GCP object,
}
