% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tin_methods.R
\name{summary.hybrid_affine_tin_fit}
\alias{summary.hybrid_affine_tin_fit}
\title{Summary Method for hybrid_affine_tin_fit Objects}
\usage{
\method{summary}{hybrid_affine_tin_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{hybrid_affine_tin_fit}.}

\item{...}{Additional arguments (not used).}
}
\description{
Summary Method for hybrid_affine_tin_fit Objects
}
