% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tin_methods.R
\name{summary.tin_akima_fit}
\alias{summary.tin_akima_fit}
\title{Summary Method for tin_akima_fit Objects}
\usage{
\method{summary}{tin_akima_fit}(object, ...)
}
\arguments{
\item{object}{An object of class \code{tin_akima_fit}.}

\item{...}{Additional arguments (not used).}
}
\description{
Summary Method for tin_akima_fit Objects
}
