% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pai_model.R
\name{surface}
\alias{surface}
\title{Plot Surface Method for the package's Objects}
\usage{
surface(object, n_grid, ...)
}
\arguments{
\item{object}{An object from the \code{mapAI} package.}

\item{...}{Additional arguments.}
}
\value{
A surface representation of the object.
}
\description{
This is a generic function that generates a surface
from the package's objects
}
\examples{
# See ?train_pai_model for a complete, runnable example.
}
