% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_distortion.R
\name{analyze_distortion}
\alias{analyze_distortion}
\title{Perform a Differential Distortion Analysis}
\usage{
analyze_distortion(pai_model, newdata = NULL, reference_scale = 1)
}
\arguments{
\item{pai_model}{A model object of class \code{pai_model}.}

\item{newdata}{A data frame with \code{source_x} and \code{source_y} columns.
If \code{NULL} (default),
the GCPs used to train the model will be used.}

\item{reference_scale}{A single numeric value to normalize area scale.}
}
\value{
An \code{distortion} object (a data frame) with the original points and
new columns for all calculated distortion metrics (e.g., \code{a}, \code{b},
\code{area_scale}, \code{max_shear}).
}
\description{
Computes a comprehensive set of detailed distortion metrics for
a PAI model at specified locations, based on Tissot's indicatrix theory.
}
\details{
This function is the core analytical engine of the package. It
implements a differential analysis by calculating the first partial
derivatives of the spatial transformation learned by a \code{pai_model} using
numerical differentiation. From these derivatives, it calculates key
distortion metrics that describe how shape, area, and angles are warped.

\strong{Interpreting Results by Model Type:}
The nature of the output is highly dependent on the model used.
\itemize{
\item \strong{\code{gam} & \code{tps}}: Produce a smooth, differentiable surface,
leading to spatially variable and meaningful distortion metrics.
\item \strong{\code{helmert} & \code{lm}}: Represent global transformations, resulting
in distortion metrics that are constant for every point.
\item \strong{\code{rf}}: Creates a step-like surface where local derivatives
may be zero, potentially showing no local distortion.
}
}
\examples{
  # Create data and train a model
  demo_data <- create_demo_data()
  gam_model <- train_pai_model(demo_data$gcp, method = "gam_biv")

  # Analyze distortion on the training points
  distortion_results <- analyze_distortion(gam_model)
  print(distortion_results)
  summary(distortion_results)

  # plot a selected metric
  plot(distortion_results, metric = "area_scale", diverging = TRUE)

# Plot Tissot's indicatrices using automatic scale factor
 indicatrices(distortion_results)
}
