% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_pai_model.R
\name{apply_pai_model}
\alias{apply_pai_model}
\title{Apply a PAI Model to Correct Map Coordinates}
\usage{
apply_pai_model(pai_model, map)
}
\arguments{
\item{pai_model}{A trained model object of class \code{pai_model} from \code{train_pai_model}.}

\item{map}{An \code{sf} object whose coordinates need correction.}
}
\value{
The corrected \code{sf} object with updated geometry and an optional \code{area_new} column
if the geometry contains polygons.
}
\description{
Applies a trained PAI model to correct the coordinates of an \code{sf} object.
}
