% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helmert.R
\name{helmert}
\alias{helmert}
\title{Internal Helmert 2D Transformation Solver}
\usage{
helmert(source_x, source_y, target_x, target_y)
}
\arguments{
\item{source_x}{Numeric vector of approximate ('from') x coordinates.}

\item{source_y}{Numeric vector of approximate ('from') y coordinates.}

\item{target_x}{Numeric vector of actual ('to') x coordinates.}

\item{target_y}{Numeric vector of actual ('to') y coordinates.}
}
\value{
An object of class \code{helmert} containing the calculated coefficients
and the original data centroids, ready for use by \code{predict.pai_model}.
}
\description{
Calculates the four parameters of a 2D similarity (Helmert)
transformation using a standard, numerically stable least-squares solution
based on centroids. This is an internal function called by
\code{train_pai_model}.
}
\keyword{internal}
