% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapAI-package.R
\docType{package}
\name{mapAI}
\alias{mapAI}
\alias{mapAI-package}
\title{mapAI: Positional Accuracy Improvement for Vector Maps}
\description{
The \code{mapAI} package provides a cohesive, end-to-end toolkit in R
for the Positional Accuracy Improvement (PAI) and distortion analysis of
vector maps.
}
\details{
The package is designed for researchers and practitioners in
geomatics and GIS, It provides a complete, modular workflow that guides the
user from importing data to final analysis and visualization.
\subsection{The Core Workflow}{

The main workflow follows a logical sequence:
\enumerate{
\item \strong{Data Handling:} Imports spatial data and control points using
\code{read_map()} and \code{read_gcp()}. The package also includes functions to
generate synthetic data (\code{create_demo_data()}) and save results
(\code{write_map()}).
\item \strong{Model Training & Validation:} Train a correction model using
\code{train_pai_model()} with a choice of methods \code{helmert},\code{tps}, \code{gam_biv}, \code{lm},
or a completely custom model on the fly.
\item Robustly evaluate model performance using spatial cross-validation with
\code{cv_pai_model()}.
\item \strong{Geometric Correction:} Apply the trained model to the full vector map
to get a geometrically corrected version using \code{transform_map()}.
\item \strong{Distortion Analysis & Visualization:} Go beyond correction to quantify
and understand the distortion that learns a PAI model itself.
\itemize{
\item Use \code{analyze_distortion()} to compute detailed metrics based on
Tissot's indicatrix theory (e.g., areal scale, angular distortion).
\item Use  \code{plot} to create publication-quality
visualizations of the error and distortion patterns.
}
}

The project's source code and issue tracker can be found on GitHub:
\url{https://github.com/kvantas/mapAI}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://kvantas.github.io/mapAI/}
  \item Report bugs at \url{https://github.com/kvantas/mapAI/issues}
}

}
\author{
\strong{Maintainer}: Konstantinos Vantas \email{kon.vantas@gmail.com} (\href{https://orcid.org/0000-0001-6387-8791}{ORCID})

Authors:
\itemize{
  \item Konstantinos Vantas \email{kon.vantas@gmail.com} (\href{https://orcid.org/0000-0001-6387-8791}{ORCID})
}

}
