% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcels.R
\docType{data}
\name{parcels}
\alias{parcels}
\title{Vector Map Parcels for Correction}
\format{
An \code{sf} object with 493 parcel polygon geometries and several variables:
\describe{
\item{KAK}{Numeric. A unique identifier for each parcel.}
\item{area_old}{\code{units} object. The original area of each polygon feature,
calculated by \code{read_map()}}
\item{geometry}{\code{sfc_POLYGON} or \code{sfc_MULTIPOLYGON}. The \code{sf} polygon geometry
representing the parcels in the target CRS (EPSG:2100).}
}
}
\source{
Part from the dataset of the 1925 urban cadastral map of Kastoria,
Greece as used in http://dx.doi.org/10.1111/tgis.70076
}
\usage{
parcels
}
\description{
A dataset containing vector map parcels (polygons from a cadastral map)
that are intended to be corrected using geospatial transformation methods.
}
\examples{
if (interactive()) {
  plot(sf::st_geometry(parcels))
}
}
\keyword{datasets}
