% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pai_model.R
\name{predict.pai_model}
\alias{predict.pai_model}
\title{Predict Method for pai_model Objects}
\usage{
\method{predict}{pai_model}(object, newdata, ...)
}
\arguments{
\item{object}{A trained model object of class \code{pai_model}.}

\item{newdata}{A data frame with \code{source_x} and \code{source_y} columns.}

\item{...}{Additional arguments passed to the model's \code{predict} function.}
}
\value{
A \code{gcp} object the predicted displacement columns: \code{dx} and \code{dy} and
the predicted target coordinates \code{target_x} and \code{target_y}
}
\description{
Predicts spatial displacement vectors (dx, dy) and the target
coordinates (target_x, target_y) from a trained  \code{pai_model} object.
}
\examples{
# example code
demo_data <- create_demo_data()

# split to train and test set
set.seed(1)
indx <- sample(1:nrow(demo_data$gcp), 100)
train_set <- demo_data$gcp[indx, ]
test_set <- demo_data$gcp[indx, ]

# fit a linear model
lm_model <- train_pai_model(train_set, method = "lm")

# predict using the test_set
pred_gcp <- predict(lm_model, test_set)
pred_gcp
}
