% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_distortion.R
\name{print.distortion}
\alias{print.distortion}
\title{Print Method for distortion Objects}
\usage{
\method{print}{distortion}(object, ...)
}
\arguments{
\item{...}{Additional arguments (not used).}

\item{x}{An object of class \code{distortion}.}
}
\description{
Custom print method for \code{distortion} objects to provide a concise
summary of the distortion analysis results.
}
\examples{
# See ?analyze_distortion for a runnable example that creates a
# distortion object.
}
