% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_map.R
\name{read_map}
\alias{read_map}
\title{Read a Vector Map for Correction}
\usage{
read_map(shp_path, ...)
}
\arguments{
\item{shp_path}{A character string specifying the path to the input map file
(e.g., a shapefile).}

\item{...}{Arguments passed to \code{sf::st_read}}
}
\value{
An \code{sf} object of the map to be corrected.
}
\description{
Reads a vector map (e.g., shapefile) that is intended to be
corrected.
}
\details{
This function reads the geospatial file using \code{sf::st_read}. If the
file contains polygon geometries and does not already have an area column
named 'area_old', this function will calculate the area of each feature and
add it.
}
\examples{
\dontrun{
# First, create a demo shapefile
demo_files <- create_demo_data()

# Read the map that needs correction
map_to_correct <- read_map(shp_path = demo_files$shp_path, quiet = TRUE)

# The output is now ready to be used in the `correct_map()` function
# after a model has been trained.
plot(sf::st_geometry(map_to_correct))
}
}
