% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pai_model.R
\name{residuals.pai_model}
\alias{residuals.pai_model}
\title{Residuals Method for pai_model Objects}
\usage{
\method{residuals}{pai_model}(x)
}
\arguments{
\item{x}{An object of class \code{pai_model}.}
}
\description{
Residuals Method for pai_model Objects
}
\details{
The residuals method computes the residuals of the model and returns
a list with them and two plots for dx and dy.
}
