% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pai_model.R
\name{surface}
\alias{surface}
\title{Create distortion surfaces from a PAI Model}
\usage{
surface(object, ...)
}
\arguments{
\item{object}{The fitted model object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
A surface representation (e.g. a list with two ggplots).
}
\description{
This is a generic function that generates a surface
representation for both \code{dx} and \code{dy}  from a fitted model object. T
}
\examples{
# See ?train_pai_model for a complete, runnable example.
}
