% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pai_model.R
\name{surface}
\alias{surface}
\title{Plot Surface Method for the package's Objects}
\usage{
surface(object, n_grid, plot_gcp, dx_range, dy_range)
}
\arguments{
\item{object}{An object from the \code{mapAI} package.}

\item{n_grid}{The resolution of the interpolation grid used to create the
smooth surface. Higher values create a more detailed plot but take longer to
compute.}

\item{plot_gcp}{A logical value indicating whether to plot the GCP
locations on the correction surfaces.}

\item{dx_range}{A numeric vector of length 2 specifying the limits for the
\code{dx} color scale.}

\item{dy_range}{A numeric vector of length 2 specifying the limits for the
\code{dy} color scale.}
}
\value{
A surface representation (e.g. a list with two ggplots).
}
\description{
This is a generic function that generates a surface
from the package's objects
}
\examples{
# See ?train_pai_model for a complete, runnable example.
}
