% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_pai_model.R
\name{train_pai_model}
\alias{train_pai_model}
\title{Train a PAI Model}
\usage{
train_pai_model(gcp_data, method, seed = 123, ...)
}
\arguments{
\item{gcp_data}{An \code{sf} object of homologous points from \code{read_gcps()}.}

\item{method}{A character string specifying the algorithm. One of:
"lm", "gam", "rf".}

\item{seed}{An integer for setting the random seed for reproducibility.}

\item{...}{Additional arguments passed to the underlying model fitting
functions (\code{mgcv::gam}, \code{stats::lm}, \code{ranger::ranger}).}
}
\value{
A trained model object of class \code{pai_model}.
}
\description{
Trains a supervised learning model to predict spatial
displacements using random forests, linear regression, or bivariate GAMs.
}
